+++ application/config/email.php
@@ -6,20 +6,20 @@
 | -------------------------------------------------------------------
 */
 
-$config["useragent"] = "CodeIgniter";
+$config["useragent"] = "trackr.moe";
 
 $config["protocol"] = "smtp";
 //$config["mailpath"] = "/usr/sbin/sendmail";
 
 
 //SMTP SETTINGS ARE SET IN SUB FOLDERS.
-$config['smtp_host']      = "smtp.zoho.com";
-$config['smtp_port']      = "465";
-$config["smtp_user"]      = "";
+$config['smtp_host']      = "smtp.mailgun.org";
+$config['smtp_port']      = "587";
+$config["smtp_user"]      = "postmaster@mg.trackr.moe";
 $config["smtp_pass"]      = "";
 $config["smtp_timeout"]   = 10;
 $config["smtp_keepalive"] = FALSE;
-$config["smtp_crypto"]    = "ssl";
+$config["smtp_crypto"]    = "tls";
 
 $config["wordwrap"]  = TRUE;
 $config["wrapchars"] = 76;