+++ public/CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [2018-01-19]
+### Changed
+- Moved from Zoho to Mailgun for our email provider.
+
 ### Fixed
 - MangaDex works again.
   - This may continue to break as the site gets updated, hopefully they'll avoid major HTML changes...